Agents/requirements.txt
# DevOps Agents System Requirements

# Core dependencies
aiohttp>=3.8.0
aiofiles>=23.0.0
psutil>=5.9.0
pyyaml>=6.0
python-dotenv>=1.0.0

# AI and ML
openai>=1.0.0
langchain>=0.1.0

# Monitoring and metrics
prometheus-client>=0.19.0
influxdb-client>=1.38.0

# File operations
watchdog>=3.0.0
pathlib>=1.0.0

# Utilities
click>=8.1.0
rich>=13.0.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0